requirements.txt
runpod
requests
Pillow
numpy<2.0.0
onnxruntime-gpu==1.18.1
rembg
fastapi
uvicorn
